pyproject.toml
[project]
name = "localllm-tgbot"
version = "0.1.0"
description = "Telegram LLM Bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "aiofiles>=25.1.0",
    "aiogram>=3.26.0",
    "diffusers>=0.37.0",
    "httpx==0.26.0",
    "langchain==0.2.16",
    "langchain-openai==0.1.22",
    "omegaconf>=2.3.0",
    "openai-whisper>=20250625",
    "python-dotenv==1.0.1",
    "python-telegram-bot==20.8",
    "soundfile>=0.12.1",
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    "transformers>=5.3.0",
]
